Changed the design behind the help button a little bit.
Now, when it is clicked, it simply brings up the help dialog. The
constructor of the help dialog simply selects the first item in the help
tree widget.

If we want to include a welcome page, we could do so by adding another
item to that widget, and remove it after it's been selected.