Highlighting works now! I changed the look and feel of the app a little bit (labels are no longer in raised panels) to see how it would look, and I think it's rather aesthetic.
Anyhow, see how you like it, we can always change it back.

Also changed the font size of the memory dump pane to be smaller.

One thing I noticed about the memory dump pane is that we ought to save
the position of the scroll bar, and make it so it doesn't go to the top
after a refresh.